# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /artists:
    get:
      operationId: listArtists
      tags: [artists]
      summary: Search artists
      parameters:
        - in: query
          name: genre
          schema:
            type: string
        - in: query
          name: minRate
          schema:
            type: integer
        - in: query
          name: maxRate
          schema:
            type: integer
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: eventDate
          schema:
            type: string
            format: date-time
          description: Filter to artists available on this date (not blocked)
        - in: query
          name: limit
          schema:
            type: integer
      responses:
        "200":
          description: Artist search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtistList"
  /artists/{id}:
    get:
      operationId: getArtist
      tags: [artists]
      summary: Get public artist profile
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Artist profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtistPublicProfile"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /artists/me:
    get:
      operationId: getArtistMe
      tags: [artists]
      summary: Get current artist public profile
      responses:
        "200":
          description: Artist profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtistPublicProfile"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    patch:
      operationId: updateArtistMe
      tags: [artists]
      summary: Update current artist profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArtistOnboardingInput"
      responses:
        "200":
          description: Updated artist profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtistProfile"
        "400":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /artists/me/gigs:
    get:
      operationId: listArtistGigs
      tags: [artists]
      summary: List upcoming gigs for the current artist (confirmed lineup + pending invites)
      security:
        - cookieAuth: []
      responses:
        "200":
          description: Upcoming gigs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ArtistGigList"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
        "404":
          description: Profile not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /artists/me/blocked-dates:
    get:
      operationId: listBlockedDates
      tags: [artists]
      summary: List blocked availability dates
      responses:
        "200":
          description: Blocked dates
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BlockedDateList"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
    post:
      operationId: addBlockedDate
      tags: [artists]
      summary: Block a date from availability
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddBlockedDateInput"
      responses:
        "201":
          description: Date blocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AddBlockedDateResult"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
  /artists/me/blocked-dates/{date}:
    delete:
      operationId: removeBlockedDate
      tags: [artists]
      summary: Remove a blocked date
      parameters:
        - in: path
          name: date
          required: true
          schema:
            type: string
            format: date
      responses:
        "204":
          description: Removed
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiError"
...
components:
  schemas:
...
    ApiError:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    ArtistOnboardingInput:
      type: object
      required: [displayName, genres, vibes]
      properties:
        displayName:
          type: string
          minLength: 1
          maxLength: 40
        avatarUrl:
          type: string
        bio:
          type: string
        genres:
          type: array
          items:
            type: string
        vibes:
          type: array
          items:
            type: string
        spotifyUrl:
          type: string
        youtubeUrl:
          type: string
...
    ArtistProfile:
      type: object
      required: [id, displayName]
      properties:
        id:
          type: integer
        displayName:
          type: string
        bio:
          type: ["string", "null"]
        genres:
          type: array
          items:
            type: string
        vibes:
          type: array
          items:
            type: string
        spotifyUrl:
          type: ["string", "null"]
        youtubeUrl:
          type: ["string", "null"]
...
    ArtistSearchResult:
      type: object
      required: [id, displayName]
      properties:
        id:
          type: integer
        displayName:
          type: string
        bio:
          type: ["string", "null"]
        genres:
          type: array
          items:
            type: string
        vibes:
          type: array
          items:
            type: string
        spotifyUrl:
          type: ["string", "null"]
        youtubeUrl:
          type: ["string", "null"]
        rateTier:
          type: ["integer", "null"]
        avatarUrl:
          type: ["string", "null"]
        username:
          type: string
    ArtistList:
      type: object
      required: [artists]
      properties:
        artists:
          type: array
          items:
            $ref: "#/components/schemas/ArtistSearchResult"
    ArtistGigSummary:
      type: object
      required: [eventId, title, eventDate, eventStatus, gigStatus, venue]
      properties:
        eventId:
          type: integer
        outreachId:
          type: ["integer", "null"]
          description: Set when gigStatus is pending (for chat accept flow)
        title:
          type: string
        eventDate:
          type: string
          format: date-time
        eventStatus:
          type: string
          description: Event planning state (upcoming or finalized)
        gigStatus:
          type: string
          enum: [confirmed, pending]
        venue:
          type: object
          required: [id, name, address]
          properties:
            id:
              type: integer
            name:
              type: string
            address:
              type: string
            ownerUsername:
              type: string
    ArtistGigList:
      type: object
      required: [gigs]
      properties:
        gigs:
          type: array
          items:
            $ref: "#/components/schemas/ArtistGigSummary"
    VenuePlayedSummary:
      type: object
      required: [id, name, address]
      properties:
        id:
          type: integer
        name:
          type: string
        address:
          type: string
    ArtistPublicProfile:
      type: object
      required: [id, displayName]
      properties:
        id:
          type: integer
        displayName:
          type: string
        bio:
          type: ["string", "null"]
        genres:
          type: array
          items:
            type: string
        vibes:
          type: array
          items:
            type: string
        spotifyUrl:
          type: ["string", "null"]
        youtubeUrl:
          type: ["string", "null"]
        rateTier:
          type: ["integer", "null"]
        avatarUrl:
          type: ["string", "null"]
        followerCount:
          type: integer
        venuesPlayed:
          type: array
          items:
            $ref: "#/components/schemas/VenuePlayedSummary"
        ratingAverage:
          type: number
        ratingCount:
          type: integer
    BlockedDateList:
      type: object
      required: [dates]
      properties:
        dates:
          type: array
          items:
            type: string
            format: date
    AddBlockedDateInput:
      type: object
      required: [date]
      properties:
        date:
          type: string
          format: date
    AddBlockedDateResult:
      type: object
      required: [date]
      properties:
        date:
          type: string
          format: date
...
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: connect.sid