# lib/api-zod/src/generated/types/artistGigList.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ArtistGigSummary } from './artistGigSummary';

export interface ArtistGigList {
  gigs: ArtistGigSummary[];
}